refactor: Move cache to root level
