+++ packages/server/src/internals/server/index.ts
@@ -4,9 +4,9 @@ import { Pool, createPool } from 'mysql2/promise'
 
 import { applyGraphQLMiddleware } from './graphql-middleware'
 import { applySwrQueueDashboardMiddleware } from './swr-queue-dashboard-middleware'
-import { createCache, createEmptyCache } from '../cache'
 import { initializeSentry } from '../sentry'
 import { createSwrQueue } from '../swr-queue'
+import { createCache, createEmptyCache } from '~/cache'
 import { createAuthServices, AuthServices } from '~/context/auth-services'
 import { Cache } from '~/context/cache'
 import { SwrQueue } from '~/context/swr-queue'
